Fixed typo causing #60 and caused by #50
I have no idea  how that crept into a *copy and paste* but my guess is that that's the reason for #60 (at least it behaves exactly the same when I change my haproxy config to that). Sorry for that, again, I have no idea how that happened, I copy pasted the config.