Allows network configuration without booting
/etc/network/interfaces file is modified such that it includes
a new file /boot/octopi-network.txt which contains sample config
snippets for getting wifi connectivity configured for WPA/WPA2-PSK,
WEP and unencrypted and also for a static IP on a virtual second
ethernet port.

This way after flashing the image to the SD the user can just
edit that file (which is on the root of the FAT partition which
desktop OSs should usually offer to access just when plugging in
the card), uncomment the matching configuration snippet, fill
in the credentials (SSID and password, or just the SSID in case
of an unencrypted WiFi), save, put the card in the Pi's SD card
slot and their USB WiFi dongle in a USB port, boot and have their
Pi directly on their network. No more cable needed, no problem
to quickly change the config when being on the move with the Pi.

Note that I tested this with the new 15-05-05 Raspbian image (that
includes that dhcpcd stuff), hence the "manual" mode on the wifi
config segments instead of "dhcp".

Test by flashing, editing the octopi-network.txt according to my
local WiFi setup, booting and verifying I could immediately connect.

Next step: netconnectd for everyone ;)