Suppress error when using restart command
When issuing one of the restart/reboot/shutdown commands
from the system menu, the request could not be acknowledged
anymore because the server was already gone, leading to the
client displaying an error for the command execution just
before the offline overlay.

To stop things like this from happening, 1.2.0 introduced
two new parameters for system actions, asnyc and ignore.

async makes the request return without waiting for the
result of the command's execution. ignore makes the client
ignore any errors in the command's execution.

The combination of both yields successful restarts/reboots/
shutdowns without scary messages.