Added mapper for network configuration
Will decide between OctoPi and Raspbian type of network
configuration depending on whether the user set up a network
configuration in /boot/octopi-network.txt or not.

A mapping script will try to see if it can find an uncommented
line

  iface $IFACE-octopi

in /boot/octopi-network.txt. If so, the user obviously used this
method to setup their network credentials (the lines are usually
all commented out) and $IFACE will be mapped to use the
$IFACE-octopi definition from that file. If no such line is
found, $IFACE will be mapped to $IFACE-raspbian, which is the
default wlan0/wlan1 configuration shipped with vanilla Raspbian.

See also #122